Seperated out Starboard checker into 2 objects that check for add event and remove event respectively.
