+++ src/main/eventhandler/MessageReactionAddEventHandler.ts
@@ -1,8 +1,8 @@
 import { MessageReaction } from 'discord.js';
 import { EventHandler } from './EventHandler';
-import { StarboardChecker } from '../modules/starboard/StarboardChecker';
 import { StarboardResponse } from '../modules/starboard/StarboardResponse';
 import { Storage } from '../storage/Storage';
+import { StarboardAddReactChecker } from '../modules/starboard/StarboardChecker/StarboardAddReactChecker';
 
 export class MessageReactionAddEventHandler extends EventHandler {
     public static EVENT_NAME = 'messageReactionAdd';
@@ -23,10 +23,14 @@ export class MessageReactionAddEventHandler extends EventHandler {
     public async handleEvent(): Promise<void> {
         const server = this.getServer(this.reaction.message.guild.id.toString());
         const { starboardSettings } = server;
-        const starboardChecker = new StarboardChecker(starboardSettings, this.reaction);
+        const starboardChecker = new StarboardAddReactChecker(starboardSettings, this.reaction);
+
+        // Check if the react qualifies to make changes.
         const shouldMakeChanges = await starboardChecker.checkAddReact();
         if (shouldMakeChanges) {
             const starboardResponse = new StarboardResponse(starboardSettings, this.reaction);
+
+            // If message exists in starboard channel, edit the count, else add to starboard
             const exists = await starboardChecker.checkIfMessageExists();
             if (exists) {
                 await starboardResponse.editStarboardMessageCount(
